final/local/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# Snowflake
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.1

# Google Gemini
google-generativeai==0.3.1

# Integrations
jira==3.5.2
datadog==0.47.0

# Utilities
httpx>=0.25.1
typing-extensions>=4.8.0

# WebSocket
websockets==12.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
